SONARIAC-813 Rule S6364: ARM Defining a short backup retention duration is security-sensitive (JSON) - Microsoft.RecoveryServices/vaults (#753)

+++ iac-extensions/arm/src/main/java/org/sonar/iac/arm/checks/LogRetentionCheck.java
@@ -91,7 +91,7 @@ public class LogRetentionCheck extends AbstractArmResourceCheck {
       if (!expr.is(ArmTree.Kind.NUMERIC_LITERAL)) {
         return false;
       }
-      float retentionDays = ((NumericLiteral) expr).value();
+      double retentionDays = ((NumericLiteral) expr).value();
       return retentionDays < retentionPeriodInDays && retentionDays != 0;
     };
   }
